37 filter on min max vals (#80)

+++ frontend/src/interfaces/index.ts
@@ -2,7 +2,7 @@ export * from "./backend";
 export * from "./mantine";
 
 export enum KVType {
-	NUMBER = "integer",
+	NUMBER = "float",
 	STRING = "string",
 }
 export interface IAttrKey {
